feat:   新增完整分帳表的檢視畫面   新增查詢行程中所有分帳表的功能 modify:   重構燈箱畫面風格

+++ IdentityJWT/Services/WeatherService.cs
@@ -15,7 +15,7 @@ namespace PinPinServer.Services
         }
 
         //呼叫獲取天氣資料的API
-        public async Task<string> GetWeatherData(string units, double lat, double lon)
+        public async Task<string> GetWeatherData(string units, decimal lat, decimal lon)
         {
             string baseaddress = $"BaseAddress: {_httpClient.BaseAddress}";
             string weatherAPI = $"data/2.5/forecast?lat={lat}&lon={lon}&appid={_apiKey}&units={units}";
